reorganization of the code: the complex data of an item are transformed into an object to facilitate access to properties (sub-data) and the evolution of sub-data.
Added functionality to have a full record instance after reading a single record and a full record collection instance after reading Many Records. A full record contains all the data on the record while a Summary record contains just the essentials, ie: the type, title and external identifiers.

Addition of validation and filter functions for the data of an orcid item with the static class: Orcid \ Work \ Data \ Data.

Addition of the possibility of requesting that they be given to be filtered by activating $ work-> setFilter () and the possibility of removing the filter with $ work-> removeFilter ()

Addition of the possibility of forcibly setting a value of a property of an object to escape the validator, which may become obsolete over time. $ work-> setPropertyByForce